+++ CHANGES.md
@@ -9,6 +9,7 @@ All notable changes to Pyink are recorded here.
   https://github.com/psf/black/pull/3640 where we still want to keep the
   parentheses around the implicitly concatenated strings if the code already
   uses them, making it more obvious it's a single function argument.
+* `--pyink-lines=` now works with stdin inputs (#16).
 
 ## 23.3.1